Add tests for step 1
